services/booking-service/requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-filter>=23.0
django-cors-headers>=4.0

# Database
psycopg2-binary>=2.9

# Validation and Serialization
pydantic>=2.0

# Message Queue (optional)
pika>=1.3  # RabbitMQ
redis>=4.5  # Redis

# HTTP Client
requests>=2.31
httpx>=0.24

# Date/Time
python-dateutil>=2.8

# Utilities
python-dotenv>=1.0

# Testing
pytest>=7.4
pytest-django>=4.5
pytest-cov>=4.1
factory-boy>=3.3

# Code Quality
black>=23.0
isort>=5.12
flake8>=6.0
mypy>=1.5

# Documentation
drf-spectacular>=0.26  # OpenAPI schema

# Monitoring
sentry-sdk>=1.29

# Production
gunicorn>=21.0
uvicorn>=0.23
